+++ CHANGES.md
@@ -77,6 +77,8 @@ v0.7 (unreleased)
   ``glue.qt.widgets.glue_mdi_area``                  | ``glue.app.qt.mdi_area``
 
   ``glue.qt.widgets.terminal``                       | ``glue.app.qt.terminal``
+  ``glue.qt.qt_roi``                                 | ``glue.core.qt.roi``
+  ``glue.core.qt.simpleforms``                            | ``glue.core.qt.simpleforms``
 
 * Improved under-the-hood creation of ROIs for Scatter and Histogram Clients. [#676]
