+++ glue/core/qt/roi.py
