+++ glue/core/qt/simpleforms.py
