+++ glue/core/qt/tests/test_simpleforms.py
