+++ glue/plugins/tools/spectrum_tool/qt/spectrum_tool.py
@@ -18,7 +18,7 @@ from glue.viewers.common.qt.toolbar import GlueToolbar
 from glue.qt.mime import LAYERS_MIME_TYPE
 from glue.qt.mouse_mode import SpectrumExtractorMode
 from glue.qt.qtutil import load_ui
-from glue.qt.simpleforms import build_form_item
+from glue.core.qt.simpleforms import build_form_item
 from glue.utils.qt.widget_properties import CurrentComboProperty
 from glue.app.qt.mdi_area import GlueMdiSubWindow
 from glue.viewers.common.qt.mpl_widget import MplWidget
